moved function definitions to Bip39.scala
